requirements.txt
streamlit
pandas
google-adk
google-genai
google-cloud-aiplatform